requirements.txt
anyio==3.6.2
appdata==2.1.2
autopep8==2.0.2
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.0.3
colorama==0.4.6
colored==1.4.3
distlib==0.3.6
dnspython==2.3.0
ecdsa==0.18.0
Faker==18.6.0
fastapi==0.95.1
fastapi-auth==0.1.1
fastapi-jwt-auth==0.5.0
filelock==3.11.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.2
numpy==1.24.2
pandas==2.0.0
passlib==1.7.4
platformdirs==3.2.0
pyasn1==0.5.0
pycodestyle==2.10.0
pydantic==1.10.7
PyJWT==1.7.1
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
requests==2.29.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
to-requirements.txt==1.3.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
virtualenv==20.21.0
